inputsource: Decimal point conversion error [1/1]
PD#TV-107892

Problem:
Failed to add, exit menu

Solution:
Decimal point conversion error

Verify:
ohm

+++ app/src/main/droidlogic/java/com/droidlogic/fragment/SatelliteWrap.java
@@ -119,7 +119,7 @@ public class SatelliteWrap {
         }
     }
 
-    public void addSatellite(String name, boolean isEast, int position) {
+    public void addSatellite(String name, boolean isEast, float position) {
         try {
             JSONArray array = new JSONArray();
             array.put(name);
